feat: user詳細画面の修正 #7
デザイン、レイアウト、表示する質問の取得するロジックの修正。タブのペジネーション、アクティブの維持実装。ファイル追加に伴い他ファイル名の変更。
