+++ .rubocop.yml
@@ -6,8 +6,7 @@ Rails:
 LineLength:
   Max: 100
   Exclude:
-    - 'app/controllers/questions_controller.rb'
-    - 'app/controllers/search_controller.rb'
+    - 'app/controllers/*'
     - 'app/uploaders/user_icon_uploader.rb'
     - 'app/uploaders/question_images_uploader.rb'
     - 'app/uploaders/answer_image_uploader.rb'
