+++ app/controllers/users_controller.rb
@@ -1,10 +1,8 @@
 class UsersController < ApplicationController
   def show
     @user = User.find(params[:id])
-    @questions = @user.questions
-    interests = Interest.where(user_id: @user.id).recent.pluck(:question_id)
-    @interesting_questions = Question.find(interests)
-    answers = Answer.where(user_id: @user.id, parent_id: nil).recent.pluck(:question_id)
-    @answered_questions = Question.find(answers)
+    @questions = @user.questions.recent.page(params[:page_post]).per(5)
+    @interesting_questions = Question.joins(:interests).where(interests: { user_id: @user.id }).order("interests.created_at desc").page(params[:page_interest]).per(5)
+    @answered_questions = Question.joins(:answers).where(answers: { user_id: @user.id, parent_id: nil }).order("answers.created_at desc").page(params[:page_answer]).per(5)
   end
 end
