--- app/javascript/src/save_acitve.js
@@ -1,14 +1,0 @@
-$(document).on('turbolinks:load', function(){
-  $(function() {
-    $('.search-sort a[data-bs-toggle="pill"]').on('click', function (e) {
-      let tabName = e.target.href;
-      let itmes = tabName.split("#");
-      Cookies.set("openTag", itmes[1], { expires: 1/12 });
-    });
-    let url = location.href;
-    if((url.match(/search/) ) && (Cookies.get("openTag"))){
-      $('a[data-bs-toggle="pill"]').removeClass('active');
-      $('a[href="#' + Cookies.get("openTag") +'"]')[0].click();
-    }
-  });
-});
